64-bit changes: ioctl cmd -> u_long, some protos.  some style, return vals.
SCCS-vsn: sys/sys/socketvar.h 8.2
SCCS-vsn: sys/sys/proc.h 8.13
SCCS-vsn: sys/sys/acct.h 8.4
SCCS-vsn: sys/sys/ucred.h 8.4
SCCS-vsn: sys/sys/mbuf.h 8.4
SCCS-vsn: sys/sys/disk.h 8.2
SCCS-vsn: sys/sys/file.h 8.3
SCCS-vsn: sys/sys/namei.h 8.5
SCCS-vsn: sys/sys/mman.h 8.2
SCCS-vsn: sys/sys/time.h 8.3
SCCS-vsn: sys/sys/cdefs.h 8.8
SCCS-vsn: sys/sys/un.h 8.2
SCCS-vsn: sys/sys/resource.h 8.4
SCCS-vsn: sys/sys/resourcevar.h 8.4
SCCS-vsn: sys/sys/signalvar.h 8.5
SCCS-vsn: sys/sys/conf.h 8.5
SCCS-vsn: sys/sys/tty.h 8.7
SCCS-vsn: sys/sys/stat.h 8.10
SCCS-vsn: sys/sys/vnode.h 8.12
SCCS-vsn: sys/sys/systm.h 8.5
SCCS-vsn: sys/sys/buf.h 8.8
SCCS-vsn: sys/sys/mount.h 8.16
SCCS-vsn: sys/sys/ioccom.h 8.3

+++ usr/src/sys/sys/disk.h
@@ -13,7 +13,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)disk.h	8.1 (Berkeley) %G%
+ *	@(#)disk.h	8.2 (Berkeley) %G%
  *
  * from: $Header: disk.h,v 1.5 92/11/19 04:33:03 torek Exp $ (LBL)
  */
@@ -46,7 +46,7 @@ struct dkdriver {
 #ifdef notyet
 	int	(*d_open) __P((dev_t dev, int ifmt, int, struct proc *));
 	int	(*d_close) __P((dev_t dev, int, int ifmt, struct proc *));
-	int	(*d_ioctl) __P((dev_t dev, int cmd, caddr_t data, int fflag,
+	int	(*d_ioctl) __P((dev_t dev, u_long cmd, caddr_t data, int fflag,
 				struct proc *));
 	int	(*d_dump) __P((dev_t));
 	void	(*d_start) __P((struct buf *, daddr_t));
